province data for China

This dataset contains spatial (sf) data for provinces in China, including various attributes related to each province.

Usage

china_province

Format

An object of class sf (inherits from data.frame) with 34 rows and 10 columns.

Source

http://datav.aliyun.com/tools/atlas/

name

The name of the province

adcode

The administrative code for the province, a unique identifier (e.g., "110000")

childrenNum

The number of administrative divisions (e.g., counties) within the province

level

The administrative level of the area, which is generally "province" for the entries in this dataset

subFeatureIndex

An index representing the sub-features within the province

centroid

The geographical centroid of the province, represented as a string of coordinates

center

The center point of the province, also represented as a string of coordinates

parent

A JSON string representing the parent administrative entity, usually the country-level data

acroutes

A JSON array of administrative codes that represent the administrative hierarchy leading to the province

geometry

Spatial geometry of the province, stored as an sf object in MULTIPOLYGON format